wokrin on email validation

+++ js/logout.mjs
@@ -5,8 +5,9 @@ logOut.addEventListener("click", checkLocalStorage)
 function checkLocalStorage(event){
     event.preventDefault();
     const getLocal = localStorage.getItem("accsessToken");
+    const getUserName = localStorage.getItem("userName");
     console.log(getLocal);
-    if (getLocal !== null){
+    if (getLocal !== null && getUserName !== null){
         window.localStorage.clear();
         window.open("signin.html", "_self");
     }
